refactor: 이벤트 생성과 관련된 서비스 및 액션 분리
